+++ packages/inquirer-common/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/inquirer-common
 
+## 0.7.5
+
+### Patch Changes
+
+-   Updated dependencies [91726b0]
+    -   @sap-ux/fiori-generator-shared@0.12.4
+
+## 0.7.4
+
+### Patch Changes
+
+-   @sap-ux/telemetry@0.6.3
+-   @sap-ux/fiori-generator-shared@0.12.3
+
 ## 0.7.3
 
 ### Patch Changes