+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sap-ux/preview-middleware
 
+## 0.20.18
+
+### Patch Changes
+
+-   920c23d: Fix custom connector for flex changes in UI5 < 1.78
+
+## 0.20.17
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.11
+
+## 0.20.16
+
+### Patch Changes
+
+-   Updated dependencies [61ea5c0]
+    -   @sap-ux/adp-tooling@0.14.10
+    -   @sap-ux/project-access@1.30.2
+    -   @sap-ux/system-access@0.6.1
+
+## 0.20.15
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.9
+
 ## 0.20.14
 
 ### Patch Changes